added await function to load data on page info boot

+++ RAAST_web/Models/Boat_Info.cs
@@ -17,7 +17,7 @@ namespace RAAST_web.Models
 
         public double? latitude { get; set; }
 
-        public float? temperature { get; set; }
+        public double? temperature { get; set; }
 
         public int? wind_Direction { get; set; }
